Refactor Pointer class to a separate file
